Support for registration of new EmitterGroup
